GAT encoder & decoders
